Fix Ant execution & stop directory creation
Summary:
Ant task 'get' executed on configuration face and unecessary directories
for checkstyle and findbugs configuration files. Now ant get executes in the begining
of its gradle task execution face and checks if its necessary to create configuration
directory files for checkstyle and findbugs.

Test Plan:
If no configuration files for findbugs and/or checkstyle are configured,
then execution should flow normally. Also if findbugs and/or checkstyle are
set to false

Reviewers: mpurita, jmsotuyo

Reviewed By: jmsotuyo

Differential Revision: http://ph.monits.com/D12518