Fix runtime error resolving configurations
Since we are now using confioguration avoidance, we may now try to
change a configuration (to resolvable) before resolution but after graph
calculation, which recent versions of Gradle don't like. So we make sure
changes to the required configurations are performed eagerly even if
configuration and resolution are kept lazy to avoid them when possible.